% Generated by roxygen2 (4.0.1): do not edit by hand
\name{biodyn}
\alias{biodyn}
\alias{biodyn,ANY-method}
\alias{biodyn-method}
\title{biodyn constructor}
\usage{
biodyn(...)
}
\arguments{
\item{...}{named parameter being passed to slots}
}
\value{
biodyn object
}
\description{
Creates an object of the biodyn class representing a biomass dynamic stock assessment model.
}
\examples{
\dontrun{
bd=biodyn(params=FLPar(r=0.6,k=50000,p=1,b0=1))
}
}

