% Generated by roxygen2 (4.0.1): do not edit by hand
\name{biodyns}
\alias{biodyns}
\alias{biodyns,biodyn-method}
\alias{biodyns,list-method}
\alias{biodyns,missing-method}
\title{biodyns}
\usage{
biodyns(object, ...)
}
\arguments{
\item{object}{biodyn object or a list of biodyn objects}

\item{...}{any other parameter}
}
\value{
biodyns object
}
\description{
Create a list with biodyn objects
}
\examples{
\dontrun{
biodyns(list("a"=bd,"b"=bd))
}
}

