% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-boot.R
\name{boot}
\alias{boot}
\alias{boot,biodyn-method}
\alias{boot-method}
\title{boot, Bootstraps biodyn.}
\usage{
boot(object, ...)
}
\arguments{
\item{object;}{a \code{biodyn} object}
}
\description{
Bootstraps the aspic model
}
\examples{
\dontrun{
    data(asp)
    asp=boot(asp)}
}
\seealso{
\code{\link{biodyn},\link{boot},\link{jk}}
}

