% Generated by roxygen2 (4.0.1): do not edit by hand
\name{control<-}
\alias{control<-}
\title{control<-}
\usage{
control(object) <- value
}
\value{
\code{biodyn} with new control slot
}
\description{
sets in \code{biodyn} initial guess and lower and upper bounds
}
\examples{
\dontrun{control(biodyn())}
}

