% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fmsy}
\alias{fmsy}
\alias{fmsy,biodyn,numeric-method}
\alias{fmsy,character,FLPar-method}
\alias{fmsy,factor,FLPar-method}
\alias{fmsy,fmsy-method,biodyn,missing-method}
\title{fmsy}
\usage{
fmsy(object, params, ...)
}
\arguments{
\item{\code{object},}{an object of class \code{biodyn} or}

\item{\code{object},}{a string or factor that species the model}

\item{\code{params},}{an \code{FLPar} object with model parameters}
}
\value{
an \code{FLPar} object with value(s) of $F_{MSY}$
}
\description{
Calculates $F_{MSY}$ given the model parameters, can be done for a biodyn class, or by specifying the model and parameters
}
\examples{
\dontrun{
fmsy('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{msy}}, \code{\link{bmsy}} and  \code{\link{refpts}}
}

