% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fwd}
\alias{fwd}
\title{fwd}
\usage{
fwd(object, ctrl, ...)
}
\arguments{
\item{\code{object},}{an object of class \code{biodyn} or  \code{biodyns}}

\item{catch}{an \code{FLQuant} or \code{FLQuants} containing future catches}

\item{harvest}{an \code{FLQuant} or \code{FLQuants} containing future harvest}

\item{stock}{an \code{FLQuant} or \code{FLQuants} containing future stock}

\item{minF}{minimum harvest, \code{FLQuant} or single numeric value}

\item{maxF}{maximum harvest, \code{FLQuant} or single numeric value}

\item{pe}{process error term, an \code{FLQuant}}

\item{peMult}{logical, i.e. by default equals true so multiplicative, otherwise additive}
}
\description{
Projects a \code{biodyn} object (i.e. a biomass dynamic model)
for a given future catch, harvest rate, or stock biomass. Only
one of these has to be supplied as an argument.
}
\examples{
\dontrun{
 bd=biodyn('logistic',FLPar(msy=1000,k=50000,b0=1),stock=FLQuant(50000,dimnames=list(year=1:50)))
 bd=fwd(bd,catch=FLQuant(500,dimnames=list(year=1:50)))
 }
}

