\name{hcrPlot}
\alias{hcrPlot}
\title{hcrPlot}
\arguments{
  \item{\code{object}}{an object of class \code{biodyn} or}

  \item{\code{params}}{\code{FLPar} object with hockey
  stock HCR parameters}
}
\value{
  a \code{FLPar} object with value(s) for HCR
}
\description{
  Calculates break pointts for a hockey stick HCR
}
\examples{
hcrPlot(object)
}
\seealso{
  \code{\link{hcr}}, \code{\link{msy}}, \code{\link{bmsy}},
  \code{\link{fmsy}} and \code{\link{refpts}}
}

