% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.biodyn}
\alias{is.biodyn}
\title{Checks class type}
\usage{
is.biodyn(x)
}
\arguments{
\item{x}{biodyn class}
}
\value{
TRUE or FALSE
}
\description{
Returns TRUE if object is of type biodyn
}
\examples{
\dontrun{
 is.biodyn(biodyn())
 }
}

