% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{kobe,biodyn,missing-method}
\alias{kobe,biodyn,missing-method}
\title{#}
\usage{
\S4method{kobe}{biodyn,missing}(object, what = c("sims", "trks", "pts",
  "smry", "wrms")[1], probs = c(0.75, 0.5, 0.25), year = NULL, nwrms = 10)
}
\description{
#
}

