% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kobe}
\alias{kobe}
\alias{kobe,biodyn,missing-method}
\alias{kobe,biodyns,missing-method}
\title{kobe}
\usage{
kobe(object, method, ...)
}
\arguments{
\item{object}{}

\item{method}{}
}
\value{
data.frame or list of data.frames
}
\description{
Creates time series of stock relative to BMSY and harvest rate relative
to FMSY
}
\examples{
\dontrun{
}
}

