% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{mvn}
\alias{mvn}
\title{mvn}
\usage{
mvn(object, n, nms = dimnames(object@control[object@control[, "phase", ] > 0,
  ])$params, fwd = FALSE, params = TRUE)
}
\arguments{
\item{object,}{a \code{biodyn} object}

\item{n,}{\code{numeric} with number of \code{iter} to create}

\item{fwd,}{\code{logical} do you want to simuate historic time series of stock biomass?, default is \code{FALSE}}
}
\value{
\code{biodyn} with simuated time series
}
\description{
Takes a fitted biodyn object and uses the covariance matrix (in the \code{vcov} slot)  and the
parameter estimates (\code{params} slot) to use Monte Carlo simulation to generate new
parameters
}
\examples{
\dontrun{
x=1
}
}

