% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,biodyn,missing-method}
\alias{plot,biodyn,missing-method}
\alias{plot,biodyns,missing}
\alias{plot,biodyns,missing-method}
\title{plot(biodyn)}
\usage{
\S4method{plot}{biodyn,missing}(x, y, probs = c(0.95, 0.5, 0.05),
  size = c(0.5, 1, 0.5), lty = c(2, 1, 2), worm = NA,
  facet = facet_wrap(~qname, scale = "free", ncol = 1), fn = list(Stock =
  function(x) stock(x), Harvest = function(x) harvest(x), Yield = function(x)
  catch(x)), ...)

\S4method{plot}{biodyns,missing}(x, y, probs = c(0.95, 0.5, 0.05),
  size = c(0.5, 1, 0.5), lty = c(2, 1, 2), facet = facet_wrap(~qname,
  scale = "free", ncol = 1), fn = list(Stock = function(x) stock(x), Harvest =
  function(x) harvest(x), Yield = function(x) catch(x)), ...)

\S4method{plot}{biodyn,missing}(x, y, probs = c(0.95, 0.5, 0.05),
  size = c(0.5, 1, 0.5), lty = c(2, 1, 2), worm = NA,
  facet = facet_wrap(~qname, scale = "free", ncol = 1), fn = list(Stock =
  function(x) stock(x), Harvest = function(x) harvest(x), Yield = function(x)
  catch(x)), ...)
}
\arguments{
\item{\code{x},}{an object of class \code{biodyn}}

\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}

\item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g. \code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a sub title for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }
}
\value{
an \code{ggplot2} object
}
\description{
Basic plot for biodyn.

plot() method for biodyns

Creates a \code{ggplot2} object that plots time series of biomass, harvest rate and catch. The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
  bd=simBiodyn()
  bd=fwd(bd,harvest=rlnorm(100,harvest(bd)[,-1]))
  plot(bd)
  }
\dontrun{
  data(bd)
  bds <- biodyns(runA=bd, runB=bd)
  plot(bds)
  }
\dontrun{
refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{plotSP}}
}

