% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotEql}
\alias{plotEql}
\alias{plotEql,biodyn,FLBRP-method}
\alias{plotEql,biodyn,FLQuant-method}
\alias{plotEql,biodyn,missing-method}
\title{plotEql}
\usage{
plotEql(x, biomass, ...)
}
\arguments{
\item{x}{an object of class \code{biodyn}}

\item{biomass}{an object of holding biomass at beginning of year}

\item{...}{other arguments}
}
\value{
an \code{ggplot2} object
}
\description{
Creates a \code{ggplot2} object that plots time series of biomass, harvest rate and catch. The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
 refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{plotPrd}}\code{\link{plotMSE}}
}

