% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotJack}
\alias{plotJack}
\title{plotJack}
\usage{
plotJack(x, y, ncol = 1)
}
\arguments{
\item{\code{x},}{an object of class \code{biodyn} that has been jack knifed, i.e. by
providing a jack knifed CPUE series to fit}

\item{\code{y}}{the original \code{biodyn} object}
}
\value{
an \code{ggplot2} object
}
\description{
Create a \code{ggplot2} plot based on a jack knifed biodyn and plots
time series of biomass and harvest rate.
The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
bd=fit(bd,cpue)
jk=fit(bd,jacknife(cpue))
plotJack(jk,bd)
}
}

