% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotSP}
\alias{plotSP}
\title{plotSP}
\usage{
plotSP(object, biomass, ...)
}
\arguments{
\item{object,}{an object of class \code{biodyn}}

\item{biomass,}{an object of holding biomass at beginning of year}
}
\value{
an \code{ggplot2} object
}
\description{
Creates a \code{ggplot2} object that plots time series of biomass, harvest rate and catch. The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
 refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{plotSP}}
}

