% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{powerAnalysis}
\alias{powerAnalysis}
\alias{powerAnalysis,biodyn,biodyn-method}
\alias{powerAnalysis,biodyn,missing-method}
\alias{powerAnalysis-method}
\title{power}
\usage{
powerAnalysis(object, ref, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{ref}{a \code{data.frame} with 3 columns \code{param,q,lower.tail} for the choice of parameter, the desired qunatile and teh tail of the distribution for the test.}

\item{...}{other arguments}
}
\value{
an \code{FLPar} object with probability of being greater (lower.tail=FALSE) or less (lower.tail=TRUE) then q
}
\description{
Estimates parameters in a \code{biodyn} class by fitting catch to CPUE indeices
}

