% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{profile,biodyn-method}
\alias{profile,biodyn-method}
\title{profile}
\usage{
\S4method{profile}{biodyn}(fitted, index, which, range = seq(0.5, 1.5,
  length.out = 21), fn = function(x) cbind(model.frame(params(x)), ll =
  model.frame(x@ll), model.frame(refpts(x))[, -4], stock = c(stock(x)[,
  ac(range(x)["maxyear"])]\%/\%bmsy(x)), harvest = c(harvest(x)[,
  ac(range(x)["maxyear"])]\%/\%fmsy(x))), run = TRUE, ...)
}
\arguments{
\item{fitted}{an object of class \code{biodyn}}

\item{index}{an \code{FLQuant}, \code{FLQuants} or  \code{data.frame} object with CPUE indices}

\item{which}{parameters to fix}

\item{range}{relative values to fix over}

\item{fn}{returned values}

\item{run}{logical, run the profile if TRUE, if FALSE set control slot and return original object}

\item{...}{additional parameters. See documentation on individual
    methods.}
}
\description{
Profiles biodyn
}
\examples{
\dontrun{
library(aspic)
library(biodyn)
data(asp)
}
}

