% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{psuedo}
\alias{psuedo}
\title{psuedo}
\usage{
psuedo(object, catch, ...)
}
\arguments{
\item{\code{object}}{of class \code{biodyn}}

\item{\code{catch}}{of class \code{FLQuant} with the catch time series.}

\item{\code{cov}}{parameter variance covariance matrix for simulation of estimation error,
by default uses the vcov slot in \code{biodyn}}
}
\value{
\code{biodyn} with estimates of stock based on catch time series
}
\description{
Simulates a \code{biodyn} assessment for a catch series, given the parameter estimates  in the \code{param}
slot and their variance covariance matrix
}
\examples{
\dontrun{
   bd=psuedo(bd,FLQuant(10000,dimnames=list(year=2009:2012)))
   plot(bd)}
}

