% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rand}
\alias{rand}
\alias{rand,biodyn,FLQuant-method}
\alias{rand-method}
\title{rand}
\usage{
rand(object, catch, ...)
}
\arguments{
\item{object}{of class \code{biodyn}}

\item{catch}{of class \code{FLQuant} with the catch time series.}

\item{...}{other arguments}
}
\value{
\code{biodyn} with estimates of stock based on catch time series
}
\description{
Simulates a \code{biodyn} object for a catch series, given the parameter estimates  in the \code{param}
slot and variance covariance matrix
}
\examples{
\dontrun{
   bd=rand(bd,FLQuant(10000,dimnames=list(year=2009:2012)))
   plot(bd)}
}

