% Generated by roxygen2 (4.0.1): do not edit by hand
\name{randJack}
\alias{randJack}
\alias{randJack,numeric,biodyn-method}
\alias{randJack-method}
\title{randJack}
\usage{
randJack(n, object)
}
\arguments{
\item{n}{\code{numeric} with number of simulations}

\item{object}{\code{biodyn}}

\item{...}{other arguments}
}
\value{
\code{biodyn} with estimates of stock based on catch time series
}
\description{
Simulates a \code{biodyn} object for a catch series, given the parameter estimates  in the \code{param}
slot and variance covariance matrix
http://young.physics.ucsc.edu/jackboot.pdf
}
\examples{
\dontrun{
bd=simBiodyn()
cpue=(stock(bd)[,-dims(bd)$year]+stock(bd)[,-1])/2
setParams(bd)=cpue
setControl(bd)=params(bd)
cpue=rlnorm(1,log(cpue),.2)
bd  =fit(bd,cpue)
bd  =fit(bd,jackknife(cpue))
bd  =randJack(100,bd)
}
}

