% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{refpts,character,FLPar-method}
\alias{refpts}
\alias{refpts,biodyn,missing-method}
\alias{refpts,character,FLPar-method}
\alias{refpts,factor,FLPar-method}
\alias{refpts-method}
\title{Maximum Sustainable Yield reference points}
\usage{
\S4method{refpts}{character,FLPar}(object = factor(object), params = params)
}
\arguments{
\item{object}{an object of class \code{biodyn} or a string or factor that species the model}

\item{params}{an \code{FLPar} object with model parameters}
}
\value{
an \code{FLPar} object with value(s) of $F_{MSY}$
}
\description{
Calculates $MSY$, $B_{MSY}$ and $F_{MSY}$ given the model parameters, can be done for a biodyn class, or by specifying the model and parameters
}
\examples{
\dontrun{
refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{msy}}, \code{\link{bmsy}}
}

