% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{controlFn}
\alias{controlFn}
\alias{runMSE}
\title{controlFn}
\usage{
controlFn(r, k, p = 1, b0 = 1, phaseR = 1, phaseK = 1, phaseP = -1,
  phaseB0 = -1, min = 0.5, max = 2)

runMSE(om, brp, srDev, uDev, range = c(min = range(om)[max], max =
  range(om)[max] + 30, interval = 3), hcr = c(ftar = 0.75, fmin = 0.01, blim =
  0.8, btrig = 0.4), ctrl = controlFn(r = calcR(brp), k = refpts(brp)["msy",
  "stock"] * 2), con = NULL, append = TRUE)
}
\arguments{
\item{om}{an \code{FLStock} object}

\item{r}{a \code{numeric} value with best guess}

\item{k}{a \code{numeric} value with best guess}

\item{p}{a \code{numeric} value with best guess, default=1}

\item{b0}{a \code{numeric} value with best guess default=1}

\item{phaseR}{a \code{numeric} value for phase, default=1}

\item{phaseK}{a \code{numeric}  value for phase, default=1,}

\item{phaseP}{a \code{numeric}  value for phase, default=-1,}

\item{phaseB0}{a \code{numeric}  value for phase, default=-1,}

\item{min}{a \code{numeric} a multipler for the best guess}

\item{max}{\code{numeric} a multipler for the best guess}

\item{om}{an \code{FLStock} objectl}

\item{brp}{an \code{FLBRP} object that holds the biological parameters for use in the projections}

\item{range}{a \code{vector} the starting and end years for the projections, and the interval for running the MP}

\item{uDev}{an \code{FLQuant} with stock recruitment residuals}

\item{hcr}{a \code{vector} with HCR options}

\item{ctrl}{a \code{vector} with HCR options}

\item{con}{a \code{vector} with HCR options}

\item{append}{a \code{vector} with HCR options}
}
\value{
a \code{control} object

invisibly a list of \code{data.frame}s with performance measures from OM and summaries from MP, if \code{con!=NULL} will
also write to a MYSQL database
}
\description{
A utility function to help set up the \code{control} slot in \code{biodyn}

Runs a full MSE using an \code{FLStock} object as the Operating Model and \code{biodyn} as the Mangement Procedure
}
\examples{
\dontrun{
   }
\dontrun{
   }
}
\seealso{
\code{\link{control}}

\code{\link{biodyn}}, \code{\link{mseBiodyn}}

\code{\link{biodyn}}, \code{\link{mseBiodyn}}
}

