% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setParams<-}
\alias{setParams<-}
\alias{setParams<-,biodyn,FLPar-method}
\alias{setParams<-,biodyn,FLQuant-method}
\alias{setParams<-,biodyn,FLQuants-method}
\alias{setParams<-,biodyn,data.frame-method}
\title{setParams<-}
\usage{
setParams(object) <- value
}
\arguments{
\item{\code{object},}{an object of class \code{biodyn}}
}
\description{
Sets up the param slot in a biodyn object given an index
}
\examples{
\dontrun{
data(bd)
setParams(bd) <-swonIndex
setParams(bd)}
}

