% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim}
\alias{sim}
\title{sim}
\usage{
sim(model = "pellat", params = FLPar(r = 0.5, k = 1000, p = 1, b0 =
  1), harvest = FLQuant(FLQuant(c(seq(0, 1.5, length.out = 30), rev(seq(0.5,
  1.5, length.out = 15))[-1], rep(0.5, 5))) * fmsy(model, params)),
  bounds = c(0.1, 10), ...)
}
\arguments{
\item{model}{character corresponding to model}

\item{params}{surplus production parameters}

\item{harvest}{\code{FLQuant} with harvest rate}

\item{bounds}{on \code{control}}

\item{...}{other arguments}
}
\value{
biodyn object with simulated time series
}
\description{
Creates a biodyn object with known properties
}
\examples{
\dontrun{
 bd=sim()
 }
}

