% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{tac,biodyn-method}
\alias{tac}
\alias{tac,biodyn-method}
\alias{tac-method}
\title{tac ,}
\usage{
\S4method{tac}{biodyn}(object, harvest, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{harvest}{an \code{FLQuant} object with harvest rate}

\item{...}{other arguments}
}
\value{
FLQuant object with TAC value(s)
}
\description{
Calculates the Total Allowable Catch for a given harvest rate and stock biomass
}
\examples{
\dontrun{
tac('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{hcr}},  \code{\link{fwd}}
}

