% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{tseries}
\alias{tseries}
\title{tseries}
\usage{
tseries(object, refpts, ...)
}
\arguments{
\item{stk}{\code{object}}

\item{refpts}{\code{FLBRP}}

\item{proxy}{\code{character} with name of reference point in refpts(brp), by default 'msy'.}
}
\value{
a \code{data.frame}
}
\description{
A utility method to create a data.frame with time series of Performance Measures i.e.
stock biomass, SSB, recruitment, F, harvest rate
}
\examples{
\dontrun{tseries(ple4,FLBRP(ple4))}
}
\seealso{
\code{\link{kobe}}, \code{\link{mseBiodyn}}
}

