% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{biodyn-class}
\alias{biodyn,character,FLPar-method}
\alias{biodyn,character,missing-method}
\alias{biodyn,factor,FLPar-method}
\alias{biodyn,factor,missing-method}
\alias{biodyn,missing,missing-method}
\alias{biodyn-class}
\title{biodyn class}
\value{
biodyn object
}
\description{
A class to represent a biomass dynamic stock assessment model.
There are slots for data, parameters and residuals and methods for calculating
reference points and other derived quantities.
}
\examples{
\dontrun{biodyn()}
}

