% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-constructors.R
\name{biodyn}
\alias{biodyn}
\alias{biodyn,ANY,ANY-method}
\alias{biodyn,ANY-method}
\alias{biodyn,FLBRP,FLStock-method}
\alias{biodyn-method}
\title{biodyn constructor}
\usage{
biodyn(object, y, ...)
}
\arguments{
\item{...}{named parameter being passed to slots}
}
\value{
biodyn object
}
\description{
Creates an object of the biodyn class representing a biomass dynamic stock assessment model.
}
\examples{
\dontrun{
bd=biodyn(params=FLPar(r=0.6,k=50000,p=1,b0=1))
}
}

