% Generated by roxygen2 (4.0.1): do not edit by hand
\name{biodyn}
\alias{biodyn}
\title{biodyn Class}
\usage{
biodyn(model, params, ...)
}
\arguments{
\item{model}{a factor or string that specifies the model type, has to be one of 'fox', 'schaefer', 'pellat', 'gulland', 'fletcher', 'shepherd', 'logistic', 'genfit'}

\item{params}{model parameters}
}
\value{
biodyn object
}
\description{
Creates an object of the biodyn class representing a biomass dynamic stock assessment model.
}
\examples{
\dontrun{
bd=biodyn('pellat',FLPar(r=0.6,k=50000,p=1,b0=1))
}
}

