% Generated by roxygen2 (4.0.1): do not edit by hand
\name{biodyns}
\alias{biodyns}
\alias{biodyns,biodyn-method}
\alias{biodyns,list-method}
\alias{biodyns,missing-method}
\title{biodyns}
\usage{
biodyns(object, ...)
}
\arguments{
\item{object}{\code{biodyn} object or a \code{list} of \code{biodyn} objects}

\item{...}{additional \code{biodyn} objects}
}
\value{
\code{biodyns} object
}
\description{
Create a list with biodyn objects
}
\examples{
\dontrun{
biodyns(biodyn())
}
}

