% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-accessors.R
\name{control}
\alias{control}
\title{control}
\usage{
control(object, method, ...)
}
\value{
FLPar
}
\description{
sets initial guess and lower and upper bounds
}
\examples{
\dontrun{control(biodyn())}
}

