% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit}
\alias{fit}
\alias{fit,biodyn,FLQuant-method}
\alias{fit,biodyn,FLQuantJK-method}
\alias{fit,biodyn,FLQuants-method}
\alias{fit-method}
\title{fit}
\usage{
fit(object, index, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{index}{an \code{FLQuant}, \code{FLQuants} or  \code{data.frame} object with CPUE indices}

\item{...}{other arguments}
}
\description{
Estimates parameters in a \code{biodyn} class by fitting catch to CPUE indices
}

