% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fwd}
\alias{fwd}
\alias{fwd,biodyn,FLQuants-method}
\alias{fwd,biodyn,missing-method}
\alias{fwd-method}
\title{fwd}
\usage{
fwd(object, ctrl, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or  \code{biodyns}}

\item{ctrl}{missing}

\item{...}{any other parameters}
}
\description{
Projects a \code{biodyn} object (i.e. a biomass dynamic model)
for a given future catch, harvest rate, or stock biomass. Only
one of these has to be supplied as an argument.
}
\examples{
\dontrun{
 bd=biodyn('logistic',FLPar(msy=1000,k=50000,b0=1),stock=FLQuant(50000,dimnames=list(year=1:50)))
 bd=fwd(bd,catch=FLQuant(500,dimnames=list(year=1:50)))
 }
}

