% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr}
\alias{hcr}
\alias{hcr,biodyn-method}
\title{hcr}
\usage{
hcr(object, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{...}{other parameters, i.e.
params \code{FLPar} object with hockey stick HCR parameters, see hcrParam
yrs numeric vector with yrs for HCR prediction
refYrs numeric vector with years used to for stock/ssb in HCR
tac \code{logical} should return value be TAC rather than F?
bndF \code{vector} with bounds (i.e.min and max values) on iter-annual variability on  F
bndTac \code{vector} with bounds (i.e. min and max values) on iter-annual variability on TAC}
}
\value{
\code{FLPar} object with value(s) for F or TAC if tac==TRUE
}
\description{
Harvest Control Rule, calculates F, or Total Allowable Catch (TAC) based on a hockey stock harvest control rule.
}
\examples{
\dontrun{
bd   =simn()

bd=window(bd,end=29)
for (i in seq(29,49,1))
bd=fwd(bd,harvest=hcr(bd,refYrs=i,yrs=i+1)$hvt)
}
}
\seealso{
\code{\link{bmsy}}, \code{\link{fmsy}}, \code{\link{fwd}} and \code{\link{hcrParam}}
}

