% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{hcrPlot,biodyn-method}
\alias{hcrPlot,biodyn-method}
\alias{hcrPlot-method}
\title{hcrPlot}
\usage{
\S4method{hcrPlot}{biodyn}(object, params = FLPar(ftar = 0.7, btrig = 0.7,
  fmin = 0.01, blim = 0.2), maxB = 1, rel = TRUE)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{params}{\code{FLPar} object with hockey stock HCR parameters}

\item{maxB}{=1}

\item{rel}{=TRUE}
}
\value{
a \code{FLPar} object with value(s) for HCR
}
\description{
Calculates break pointts for a hockey stick HCR
}
\examples{
\dontrun{
sim()
}
}
\seealso{
\code{\link{hcr}},  \code{\link{msy}},  \code{\link{bmsy}}, \code{\link{fmsy}}
}

