% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{hcrPlot,biodyn-method}
\alias{hcrPlot,biodyn-method}
\title{hcrPlot}
\usage{
\S4method{hcrPlot}{biodyn}(object, params = FLPar(ftar = 0.7, btrig = 0.7,
  fmin = 0.01, blim = 0.2), maxB = 1, rel = TRUE)
}
\arguments{
\item{\code{object}}{an object of class \code{biodyn} or}

\item{\code{params}}{\code{FLPar} object with hockey stock HCR parameters}
}
\value{
a \code{FLPar} object with value(s) for HCR
}
\description{
Calculates break pointts for a hockey stick HCR
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{hcr}},  \code{\link{msy}},  \code{\link{bmsy}}, \code{\link{fmsy}} and  \code{\link{refpts}}
}

