% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,biodyn,missing-method}
\alias{plot,biodyn,missing-method}
\alias{plot,biodyns,missing-method}
\title{plot}
\usage{
\S4method{plot}{biodyn,missing}(x, y, probs = c(0.95, 0.75, 0.5, 0.25, 0.05),
  na.rm = FALSE, type = 7, worm = NULL, fn = list(Stock = function(x)
  stock(x), Harvest = function(x) harvest(x), Yield = function(x) catch(x)),
  facet = facet_wrap(~qname, scales = "free", ncol = 1), ...)
}
\arguments{
\item{x}{an object of class \code{biodyn}}

\item{worm}{iters}

\item{y}{second argument}

\item{probs}{numeric vector of probabilities with values in [0,1]. (Values up to 2e-14 outside that range are accepted and moved to the nearby endpoint.)}

\item{na.rm}{FALSE}

\item{type}{an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used.}

\item{fn}{functions}

\item{facet}{facet for panels}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g. \code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a sub title for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }
}
\value{
an \code{ggplot2} object
}
\description{
Creates a \code{ggplot2} object that plots time series of biomass, harvest rate and catch. The basic object can then be modified by adding ggplot2 layers.
}
\examples{
\dontrun{

x =simBiodyn()
x =window(x,end=49)
bd1=fwd(x,harvest=rlnorm(200,log(harvest(x)[,-1]),.2))
bd2=fwd(x,harvest=rlnorm(200,log(harvest(x)[,-1])*1.5,.2))
plot(biodyns("1"=bd1,"2"=bd2))
}
}
\seealso{
\code{\link{plotPrd}}, \code{\link{plotEql}}
}

