% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{priorFn}
\alias{priorFn}
\title{priorFn}
\usage{
priorFn(...)
}
\arguments{
\item{...}{#'}
}
\value{
\code{list} with om, ...
}
\description{
A utility function to help set up the \code{prior} slot in \code{biodyn}.
}
\examples{
\dontrun{
   priorFn(r=c(weight=1,0.3,0.03)
   }
}
\seealso{
\code{\link{biodyn}}
}

