% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{psuedo}
\alias{psuedo}
\alias{psuedo,biodyn,FLQuant-method}
\alias{psuedo-method}
\title{psuedo}
\usage{
psuedo(object, catch, ...)
}
\arguments{
\item{object}{of class \code{biodyn}}

\item{catch}{of class \code{FLQuant} with the catch time series.}

\item{...}{other arguments}
}
\value{
\code{biodyn} with estimates of stock based on catch time series
}
\description{
Simulates a \code{biodyn} assessment for a catch series, given the parameter estimates  in the \code{param}
slot and their variance covariance matrix
}
\examples{
\dontrun{
   bd=psuedo(bd,FLQuant(10000,dimnames=list(year=2009:2012)))
   plot(bd)}
}

