% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-mse-application.R, R/biodyn-mse.R
\name{runMSE}
\alias{mse}
\alias{mseBiodyn}
\alias{runMSE}
\title{runMSE}
\usage{
runMSE(om, eql, mp, srDev, uDev, ftar = 0.75, fmin = 0.01, blim = 0.8,
  btrig = 0.4, what = "msy", mult = TRUE,
  range = c(range(om)["maxyear"], c(range(om)["maxyear"]) + 30, interval = 3))

mseBiodyn(om, eql, srDev, control, priors, start = range(om)["maxyear"],
  end = start + 30, interval = 3, uDev = 0.2, u = oem, ftar = 0.7,
  btrig = 0.6, fmin = 0.01, blim = 0.01, what = "msy", mult = TRUE,
  bndF = NULL, maxF = 1, phaseQ = 1,
  cmdOps = paste("-maxfn 500 -iprint 0 -est"), omega = 1, refB = 1,
  trendQ = 1 + FLQuant(cumprod(rep(0, end - dims(trendQ)$minyear + 1)),
  dimnames = list(year = dims(trendQ)$minyear:end)))
}
\arguments{
\item{om}{an \code{FLStock} object}

\item{eql}{an \code{FLBRP} object that holds the biological parameters for use in the projections}

\item{mp}{an \code{biodyn} object that holds the options for the biomass dynamic assessment model}

\item{srDev}{a \code{FLQuant} with recruitment deviates}

\item{uDev}{an \code{FLQuant} or \code{FLQuants} with CPUE residuals}

\item{ftar}{a \code{numeric} with target F in HCR}

\item{fmin}{a \code{numeric} with minimum F in HCR}

\item{blim}{a \code{numeric} with biomass limit for HCR}

\item{btrig}{a \code{numeric} with biomass trigger (i.e. when to reduce F) in HCR}

\item{what}{a \code{character} that specifies what is to be used for the reference point in the HCR, recycled as required}

\item{mult}{a \code{logical} that specifies whether quantity in HCR options is a multiplier or probability, recycled as required}

\item{range}{a \code{vector} the starting and end years for the projections, and the interval for running the MP}

\item{om}{an \code{FLStock} object}

\item{eql}{an \code{FLBRP} object that holds the biological parameters for use in the projections}

\item{mp}{an \code{biodyn} object that holds the options for the biomass dynamic assessment model}

\item{range}{a \code{vector} the starting and end years for the projections, and the interval for running the MP}

\item{srDev}{a \code{FLQuant} with recruitment deviates}

\item{uDev}{an \code{FLQuant} or \code{FLQuants} with CPUE residuals}

\item{ftar}{a \code{numeric} with target F in HCR}

\item{fmin}{a \code{numeric} with minimum F in HCR}

\item{blim}{a \code{numeric} with biomass limit for HCR}

\item{btrig}{a \code{numeric} with biomass trigger (i.e. when to reduce F) in HCR}

\item{what}{a \code{character} that specifies what is to be used for the reference point in the HCR, recycled as required}

\item{mult}{a \code{logical} that specifies whether quantity in HCR options is a multiplier or probability, recycled as required}
}
\value{
a list of \code{data.frame}s with performance measures from OM and summaries from MP, if \code{con!=NULL} will
also write to a MYSQL database

a list of \code{data.frame}s with performance measures from OM and summaries from MP, if \code{con!=NULL} will
also write to a MYSQL database
}
\description{
Runs a full MSE using an \code{FLStock} object as the Operating Model and \code{biodyn} as the Mangement Procedure

Runs a full MSE using an \code{FLStock} object as the Operating Model and \code{biodyn} as the Mangement Procedure
}
\examples{
\dontrun{
sim()
   }
\dontrun{
sim()
   }
}
\seealso{
\code{\link{biodyn}}

\code{\link{biodyn}}
}

