% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{runMSE}
\alias{runMSE}
\title{runMSE}
\usage{
runMSE(om, brp, srDev, uDev, range = c(min = range(om)[max], max =
  range(om)[max] + 30, interval = 3), hcr = c(ftar = 0.75, fmin = 0.01, blim =
  0.8, btrig = 0.4), ctrl = controlFn(r = calcR(brp), k = refpts(brp)["msy",
  "stock"] * 2), con = NULL, append = TRUE)
}
\arguments{
\item{om}{an \code{FLStock} objectl}

\item{brp}{an \code{FLBRP} object that holds the biological parameters for use in the projections}

\item{range}{a \code{vector} the starting and end years for the projections, and the interval for running the MP}

\item{srDev}{a \code{FLQuant} with recruitment deviates}

\item{uDev}{an \code{FLQuant} with stock recruitment residuals}

\item{hcr}{a \code{vector} with HCR options}

\item{ctrl}{a \code{vector} with HCR options}

\item{con}{a \code{vector} with HCR options}

\item{append}{a \code{vector} with HCR options}
}
\value{
invisibly a list of \code{data.frame}s with performance measures from OM and summaries from MP, if \code{con!=NULL} will
also write to a MYSQL database
}
\description{
Runs a full MSE using an \code{FLStock} object as the Operating Model and \code{biodyn} as the Mangement Procedure
}
\examples{
\dontrun{
simBiodyn()
   }
}
\seealso{
\code{\link{biodyn}}
}

