% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setControl<-}
\alias{setControl<-}
\alias{setControl<-,biodyn,FLPar-method}
\alias{setControl<-,biodyn,FLQuant-method}
\alias{setControl<-,biodyn,FLQuants-method}
\title{setControl<-}
\usage{
setControl(object) <- value
}
\arguments{
\item{\code{object},}{an object of class \code{biodyn}}
}
\description{
#' Sets up the control slot in a biodyn object given the values in the \code{params}
slot. The starting values are set to the values in \code{params} and the min and
max bounds to .1 and 10 times of these.
}
\examples{
\dontrun{
data(bd)
control(bd) <-params(bd)
params(bd)
control(bd)}
}
\seealso{
\code{\link{controlFn}}
}

