% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setParams<-}
\alias{setParams<-}
\alias{setParams<-,biodyn,FLPar-method}
\alias{setParams<-,biodyn,FLQuant-method}
\alias{setParams<-,biodyn,FLQuants-method}
\alias{setParams<-,biodyn,data.frame-method}
\title{setParams<-}
\usage{
setParams(object) <- value
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{value}{\code{params}}
}
\description{
Add parameters for catchability and sigma in the param slot given an index or indices
of stock abundance as \code{FLQuant} or \code{FLQuants}.
}
\examples{
\dontrun{
bd=simBiodyn()
cpue=stock(bd)
setParams(bd) <-cpue
setParams(bd)}
}

