% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simBiodyn}
\alias{simBiodyn}
\title{Simulates a \code{biodyn} object}
\usage{
simBiodyn(model = "pellat", params = FLPar(r = 0.5, k = 1000, p = 1, b0 = 1,
  q = 1, sigma = 0.3), harvest = FLQuant(FLQuant(c(seq(0, 1.5, length.out =
  30), rev(seq(0.5, 1.5, length.out = 15))[-1], rep(0.5, 5))) * fmsy(model,
  params)), bounds = c(0.1, 10), ...)
}
\arguments{
\item{model}{}

\item{params}{}

\item{harvest}{}

\item{bounds}{}
}
\value{
TRUE or FALSE
}
\description{
Creates a biodyn object with known properties
}
\examples{
\dontrun{
 bd=simBiodyn()
 }
}

