% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tseries}
\alias{tseries}
\alias{tseries,FLStock,missing}
\title{tseries}
\usage{
tseries(object, refpts, ...)
}
\arguments{
\item{object}{\code{biodyn}}

\item{refpts}{\code{FLBRP}}

\item{...}{other arguments}
}
\value{
a \code{data.frame}
}
\description{
A utility method to create a data.frame with time series of Performance Measures i.e.
stock biomass, SSB, recruitment, F, harvest rate
}
\examples{
\dontrun{tseries(ple4,FLBRP(ple4))}
}

